# private/protocol/json/jsonutil: reduce build allocs