--- a/private/protocol/json/jsonutil/build.go
+++ b/private/protocol/json/jsonutil/build.go
@@ -84,11 +84,12 @@ func buildStruct(value reflect.Value, buf *bytes.Buffer, tag reflect.StructTag) 
 	t := value.Type()
 	first := true
 	for i := 0; i < t.NumField(); i++ {
-		field := t.Field(i)
-		member := value.FieldByIndex(field.Index)
+		member := value.Field(i)
 		if (member.Kind() == reflect.Ptr || member.Kind() == reflect.Slice || member.Kind() == reflect.Map) && member.IsNil() {
 			continue // ignore unset fields
 		}
+
+		field := t.Field(i)
 		if field.PkgPath != "" {
 			continue // ignore unexported fields
 		}