More work on https://github.com/aws/aws-sdk-go/issues/377

Avoids fetching a field if its value is empty. There are many cases
where most fields will be empty. For example, `dynamodb.AttributeValue`.

```
benchmark                 old ns/op     new ns/op     delta
BenchmarkBuildJSON-4      8942          7807          -12.69%

benchmark                 old allocs     new allocs     delta
BenchmarkBuildJSON-4      58             45             -22.41%

benchmark                 old bytes     new bytes     delta
BenchmarkBuildJSON-4      1344          1136          -15.48%
```