@bpot 2015-11-13
Deltas for `make bench-protocol` against master:

```
benchmark                                                                      old ns/op     new ns/op     delta
BenchmarkEC2QueryBuild_Complex_ec2AuthorizeSecurityGroupEgress-4               49414         50128         +1.44%
BenchmarkEC2QueryBuild_Simple_ec2AttachNetworkInterface-4                      12972         13064         +0.71%
BenchmarkBuildJSON-4                                                           8984          7713          -14.15%
BenchmarkStdlibJSON-4                                                          7071          6934          -1.94%
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-4                                 18702         16294         -12.88%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-4                                18531         15594         -15.85%
BenchmarkEncodingJSONMarshal_Simple_dynamodbPutItem-4                          6596          6326          -4.09%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-4               273698        271139        -0.93%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-4                   8876          9031          +1.75%
BenchmarkEncodingJSONMarshal_Complex_elastictranscoderCreateJobInput-4         40054         38537         -3.79%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-4            13949         14221         +1.95%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-4                10198         9991          -2.03%
BenchmarkEncodingJSONMarshal_Simple_elastictranscoderListJobsByPipeline-4      1064          1093          +2.73%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-4                   414674        407649        -1.69%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-4           19058         18549         -2.67%
BenchmarkEncodingXMLMarshal_Simple_cloudfrontDeleteStreamingDistribution-4     5207          5172          -0.67%

benchmark                                                                      old allocs     new allocs     delta
BenchmarkEC2QueryBuild_Complex_ec2AuthorizeSecurityGroupEgress-4               237            237            +0.00%
BenchmarkEC2QueryBuild_Simple_ec2AttachNetworkInterface-4                      67             67             +0.00%
BenchmarkBuildJSON-4                                                           58             45             -22.41%
BenchmarkStdlibJSON-4                                                          13             13             +0.00%
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-4                                 105            79             -24.76%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-4                                103            77             -25.24%
BenchmarkEncodingJSONMarshal_Simple_dynamodbPutItem-4                          8              8              +0.00%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-4               1131           1131           +0.00%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-4                   59             59             +0.00%
BenchmarkEncodingJSONMarshal_Complex_elastictranscoderCreateJobInput-4         7              7              +0.00%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-4            80             80             +0.00%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-4                57             57             +0.00%
BenchmarkEncodingJSONMarshal_Simple_elastictranscoderListJobsByPipeline-4      2              2              +0.00%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-4                   1832           1832           +0.00%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-4           88             88             +0.00%
BenchmarkEncodingXMLMarshal_Simple_cloudfrontDeleteStreamingDistribution-4     9              9              +0.00%

benchmark                                                                      old bytes     new bytes     delta
BenchmarkEC2QueryBuild_Complex_ec2AuthorizeSecurityGroupEgress-4               10920         10919         -0.01%
BenchmarkEC2QueryBuild_Simple_ec2AttachNetworkInterface-4                      3696          3696          +0.00%
BenchmarkBuildJSON-4                                                           1344          1136          -15.48%
BenchmarkStdlibJSON-4                                                          1200          1200          +0.00%
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-4                                 4289          3873          -9.70%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-4                                4225          3809          -9.85%
BenchmarkEncodingJSONMarshal_Simple_dynamodbPutItem-4                          808           808           +0.00%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-4               41276         41276         +0.00%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-4                   2800          2800          +0.00%
BenchmarkEncodingJSONMarshal_Complex_elastictranscoderCreateJobInput-4         4905          4905          +0.00%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-4            4368          4368          +0.00%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-4                3328          3328          +0.00%
BenchmarkEncodingJSONMarshal_Simple_elastictranscoderListJobsByPipeline-4      120           120           +0.00%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-4                   103386        103351        -0.03%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-4           9104          9104          +0.00%
BenchmarkEncodingXMLMarshal_Simple_cloudfrontDeleteStreamingDistribution-4     4656          4656          +0.00%
```