@jasdel 2015-11-13
Hi @bpot thanks for making this improvement reducing the allocations when marshaling to JSON. The code change looks good.